# docs/api/openapi.yaml
...
openapi: 3.1.0
info:
  title: trapline API
  version: "1.0"
  summary: Errors, releases, alerts, monitors, tracing and release health for one installation.
  description: |
    One API, four clients: this REST surface, the `trapline` CLI, the web
    panel, and MCP (ADR 006). Everything the panel can do is reachable here,
    because the panel is a client of this and not a privileged insider.

    **Authentication.** Two credentials, one gate. A browser session carries a
    `SameSite=Lax` cookie and must send `X-Trapline-Request` on every
    state-changing request (the custom-header CSRF defence). A program carries
    `Authorization: Bearer ek_…` and sends no such header, because a bearer
    token is not ambient and there is nothing to forge.

    **Scopes.** Coarse on purpose. `projects:read|write` covers issues, stats,
    releases, tracing, artefacts and health. `alerts:read|write` is separate
    because an alert channel holds a credential for somebody else's system.
    `monitors:read|write` is separate because a ping key is a credential, and
    because creating an uptime monitor is the permission to make this server
    open outbound connections to an address the caller chose (ADR 016).

    **Errors.** One shape for every failure: `{"error": "…"}`, with a status
    code that means what it says. A 400 from a search is a search that could
    not run, which is not the same as a search that found nothing (ADR 011).
  license:
    name: Apache-2.0
...
servers:
...
security:
  - sessionCookie: []
  - bearerToken: []
...
paths:
...
  /projects/{id}/releases:
    parameters:
      - $ref: "#/components/parameters/ProjectID"
    get:
      tags: [releases]
      summary: Every release, newest first.
      description: |
        Ordered by semver when the versions parse as semver, and by first
        event otherwise; an unknown version sorts as newer than a known one
        (ADR 012).
      parameters:
        - $ref: "#/components/parameters/Limit"
      responses:
        "200":
          description: The list.
          content:
            application/json:
              schema:
                type: object
                properties:
                  releases:
                    type: array
                    items: { $ref: "#/components/schemas/Release" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "404": { $ref: "#/components/responses/NotFound" }
    post:
      tags: [releases]
      summary: Declare a release before its first event.
      description: |
        A release also appears on its own the first time an event names it, so
        this is for a deploy pipeline that wants to attach commits or a deploy
        before any traffic arrives.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [version]
              properties:
                version: { type: string }
                date_released: { type: string, format: date-time, nullable: true }
      responses:
        "201":
          description: The release.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Release" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
  /projects/{id}/releases/{version}:
    parameters:
      - $ref: "#/components/parameters/ProjectID"
      - $ref: "#/components/parameters/ReleaseVersion"
    get:
      tags: [releases]
      summary: One release, its commits and its deploys.
      responses:
        "200":
          description: The release.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ReleaseDetail" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "404": { $ref: "#/components/responses/NotFound" }
    put:
      tags: [releases]
      summary: Finalise a release.
      description: |
        PUT because finalising twice is the same as finalising once: the
        second call keeps the first date.
      requestBody:
        required: false
        content:
          application/json:
            schema:
              type: object
              properties:
                date_released: { type: string, format: date-time }
      responses:
        "200":
          description: The release, finalised.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Release" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
...
components:
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: trapline_session
      description: |
        The panel's credential. `SameSite=Lax`, `HttpOnly`, and `Secure` when
        the server knows it is behind TLS. Because it is ambient, every
        state-changing request must also carry `X-Trapline-Request` with any
        value; without it the request is refused with 403. A browser cannot
        set a custom header cross-origin without a preflight, and no
        permissive CORS policy is served for one to pass — which is the whole
        defence, in one header and no token plumbing.
    bearerToken:
      type: http
      scheme: bearer
      bearerFormat: ek_...
      description: |
        A program's credential, carrying scopes. Not ambient, so it needs no
        CSRF header. Scopes are `projects:read`, `projects:write`,
        `alerts:read`, `alerts:write`, `monitors:read`, `monitors:write`;
        where an operation below names none, it takes `projects:read` or
        `projects:write` according to its method.
  parameters:
    ProjectID:
      name: id
      in: path
      required: true
      schema: { type: integer, format: int64 }
      description: The project's numeric id.
...
    ReleaseVersion:
      name: version
      in: path
      required: true
      schema: { type: string }
      description: |
        Whatever a deploy tool put on an event, percent-encoded — `shop@1.4.2`
        travels as `shop%401.4.2`.
...
    Limit:
      name: limit
      in: query
      schema: { type: integer, minimum: 1 }
      description: |
        How many rows. Absent means the endpoint's default; every endpoint
        bounds its own maximum, so a large value is clamped rather than
        refused.
  responses:
    BadRequest:
      description: |
        Malformed, or a question that cannot be asked. A search that the index
        cannot run answers here and says why — which is not the same as a
        search that found nothing, and the difference is an afternoon
        (ADR 011).
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
    Unauthorized:
      description: No credential, or one that has expired or been revoked.
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
    Forbidden:
      description: |
        A valid credential without the scope this route demands — or, for a
        cookie, a state-changing request with no `X-Trapline-Request` header.
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
    NotFound:
      description: |
        No such resource. Also what a subsystem this build was assembled
        without answers: a route that does not exist is the honest description
        of a surface this installation does not have.
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
...
  schemas:
    Error:
      type: object
      description: The one failure shape, for every route in this document.
      required: [error]
      properties:
        error: { type: string }
...
    Release:
      type: object
      properties:
        version: { type: string }
        project_id: { type: integer, format: int64 }
        created_at: { type: string, format: date-time }
        date_released: { type: string, format: date-time, nullable: true }
        first_event_at: { type: string, format: date-time, nullable: true }
        last_event_at: { type: string, format: date-time, nullable: true }
        commit_count: { type: integer, format: int64 }
    CommitFile:
      type: object
      properties:
        path: { type: string }
        type: { type: string, enum: [A, M, D] }
    Commit:
      type: object
      properties:
        id: { type: string, description: The SHA. }
        message: { type: string }
        author_name: { type: string }
        author_email: { type: string }
        timestamp: { type: string, format: date-time, nullable: true }
        repository: { type: string }
        patch_set:
          type: array
          items: { $ref: "#/components/schemas/CommitFile" }
          description: |
            The files this commit touched. Without it a release still lists
            its commits, but nothing can be accused of having caused an issue
            (ADR 019).
    Deploy:
      type: object
      properties:
        id: { type: integer, format: int64 }
        environment: { type: string }
        name: { type: string }
        url: { type: string }
        started_at: { type: string, format: date-time, nullable: true }
        finished_at: { type: string, format: date-time, nullable: true }
    ReleaseDetail:
      allOf:
        - $ref: "#/components/schemas/Release"
        - type: object
          properties:
            new_issues: { type: integer, format: int64 }
            regressed_issues: { type: integer, format: int64 }
            events: { type: integer, format: int64 }
            commits:
              type: array
              items: { $ref: "#/components/schemas/Commit" }
            deploys:
              type: array
              items: { $ref: "#/components/schemas/Deploy" }
...